requirements.txt
-i https://pypi.org/simple
alembic==1.13.1; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
fastapi==0.108.0; python_version >= '3.8'
graphql-core==3.2.3; python_version >= '3.6' and python_version < '4'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
passlib==1.7.4
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
python-jose==3.3.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.24; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
strawberry-graphql==0.217.0; python_version >= '3.8' and python_version < '4.0'
typing-extensions==4.9.0; python_version >= '3.8'
uvicorn==0.25.0; python_version >= '3.8'
